Try strengthening the definition of IsNewerConfig to include terms
I do not think this will actually fix the ElectionSafety issue. If a
node has a config with an older version, but then learns of a newer
term, and then sends its older config to a node with a newer config (if
ordered by term), then we still violate config ordering. I think we may
need to store a term value with each config which represents the term
in which that config was written i.e. which primary wrote it. Then we
order configs by (term, configVersion). You reject configs if the term
of the config is older than the term of your current config.